showing info to user about reported announcment

+++ AcoPlaceBackend/src/main/java/put/poznan/AcoPlaceBackend/controller/AnnouncementController.java
@@ -24,7 +24,7 @@ public class AnnouncementController {
     }
 
     @GetMapping("/announcement/{id}")
-    public Announcement getAdvertisementById(@PathVariable long id){
+    public Announcement getAdvertisementById(@PathVariable int id){
         return announcementService.getAnnouncementById(id);
     }
